requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
typhoon-ocr==0.3.8
python-dotenv==1.0.0 